+++ lib/rubocop/cop/ascii_identifiers.rb
@@ -2,12 +2,12 @@
 
 module Rubocop
   module Cop
-    class AsciiIdentifiersAndComments < Cop
-      ERROR_MESSAGE = 'Use only ascii symbols in identifiers and comments.'
+    class AsciiIdentifiers < Cop
+      ERROR_MESSAGE = 'Use only ascii symbols in identifiers.'
 
       def inspect(file, source, tokens, sexp)
         tokens.each do |t|
-          if [:on_ident, :on_comment].include?(t.type) &&
+          if t.type == :on_ident &&
               t.text =~ /[^\x00-\x7f]/
             add_offence(:convention, t.pos.lineno, ERROR_MESSAGE)
           end