+++ lib/rubocop/cop/symbol_snake_case.rb
@@ -6,6 +6,13 @@ module Rubocop
       ERROR_MESSAGE = 'Use snake_case for symbols.'
       SNAKE_CASE = /^@?[\da-z_]+[!?=]?$/
       def inspect(file, source, tokens, sexp)
+        check_for_symbols(sexp)
+        check_for_hash_labels(sexp)
+      end
+
+      private
+
+      def check_for_symbols(sexp)
         each(:symbol_literal, sexp) do |s|
           symbol_type = s[1][1][0]
 
@@ -16,6 +23,19 @@ module Rubocop
 
           unless symbol_ident =~ SNAKE_CASE
             line_no = s[1][1][2].lineno
+            add_offence(:convention,
+                        line_no,
+                        ERROR_MESSAGE)
+          end
+        end
+      end
+
+      def check_for_hash_labels(sexp)
+        each(:@label, sexp) do |s|
+          label_ident = s[1].chop
+
+          unless label_ident =~ SNAKE_CASE
+            line_no = s[2].lineno
             add_offence(:convention,
                         line_no,
                         ERROR_MESSAGE)