+++ spec/spec_helper.rb
@@ -14,6 +14,9 @@ require 'rspec'
 require 'rubocop'
 require 'rubocop/cli'
 
+# disable colors in specs
+Sickill::Rainbow.enabled = false
+
 # Requires supporting files with custom matchers and macros, etc,
 # in ./support/ and its subdirectories.
 Dir["#{File.dirname(__FILE__)}/support/**/*.rb"].each { |f| require f }