+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 
 * [#88](https://github.com/bbatsov/rubocop/issues/88) Abort gracefully when interrupted with Ctrl-C
 * No longer crashes on bugs within cops. Now problematic checks are skipped and a message is displayed.
+* Replaced Term::ANSIColor with Rainbow.
 
 ## 0.6.0 (04/23/2013)